requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
vectorbt>=0.26.0
matplotlib>=3.7.0
